pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["ivan <ivan.ivan@karmanow.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
pydantic-settings = "^2.3.3"
uvicorn = "^0.30.1"
psycopg = { extras = ["binary"], version = "^3.1.19" }
alembic = "^1.13.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
